requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
paho-mqtt==1.4.0
pubnub==4.1.3
pycryptodomex==3.8.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
